Set the supports_order_column_alias feature.
This is mandatory for functional indexes, if the parameters are bound
server-side and Django only uses positional parameters.

For instance, with the query:

    select sum(value) from testdata
    group by left(name, %s)
    order by left(name, %s)

If the query is composed client-side, the server is able to infer that
the group by and the order by are the same expression. If the arguments
are passed to the server, they need to be the same, for instance the
following works well:

    cursor.execute(
        """
        select sum(value) from testdata
        group by left(name, %(num)s)
        order by left(name, %(num)s)
        """,
        {"num": 3})

But if only positional arguments are passed,

    cursor.execute(
        """
        select sum(value) from testdata
        group by left(name, %s)
        order by left(name, %s)
        """,
        [3, 3])

PostgreSQL will not infer that the group by and the order by are the
same expression, and will fail with

    column "testdata.name" must appear in the GROUP BY clause or be used in
    an aggregate function

If a backend is declared supports_order_column_alias, the query is
rewritten as:

    cursor.execute(
        """
        select sum(value) from testdata
        group by left(name, %s)
        order by 1
        """,
        [3])

avoiding the problem.